Code cleanup
- Fixed some warnings.
- Removed debug code that was accidentally left in a previous commit.
- Added BinaryReader::SeekReverse() which should be used instead of ::SeekCur() to move backwards. There's working code that was passing negative numbers to SeekCur(), even though it's argument is a size_t, which is unsigned. I have no idea why that code worked. May have been a lucky bug.

+++ BinaryTools/BinaryTools.cpp
@@ -51,7 +51,7 @@ int main()
             writeData.z = 22334.44f;
             writeData.cash = 1003;
             writeData.score = -64230;
-            printf("sizeof(TestPod) = %d\n", sizeof(TestPod));
+            printf("sizeof(TestPod) = %zd\n", sizeof(TestPod));
 
             printf("Writing POD struct from memory... ");
             BinaryWriter writer("./TestBin2.bin");
@@ -121,7 +121,7 @@ int main()
             printf("Reading values of memory buffer with BinaryReader... ");
             printf("Done!\n");
 
-            BinaryReader reader(span.Data(), span.Size());
+            BinaryReader reader(span.Data(), (uint32_t)span.Size());
             printf("Printing values...\n");
             printf("Float: %f\n", reader.ReadFloat());
             printf("Float: %f\n", reader.ReadFloat());
